move InOut logic to separate file
